+++ src/helpers/ImageTransforms.php
@@ -270,6 +270,16 @@ class ImageTransforms
         }
 
         if (is_array($transform)) {
+            if (!empty($transform['width']) && !is_numeric($transform['width'])) {
+                Craft::warning("Invalid transform width: {$transform['width']}", __METHOD__);
+                $transform['width'] = null;
+            }
+
+            if (!empty($transform['height']) && !is_numeric($transform['height'])) {
+                Craft::warning("Invalid transform height: {$transform['height']}", __METHOD__);
+                $transform['height'] = null;
+            }
+
             if (array_key_exists('transform', $transform)) {
                 $baseTransform = self::normalizeTransform(ArrayHelper::remove($transform, 'transform'));
                 return self::extendTransform($baseTransform, $transform);